+++ bitorch/layers/qconv2d.py
@@ -1,6 +1,6 @@
 """Module containing the quantized 2d convolution layer"""
 
-from typing import Union, Any
+from typing import Any, Type, Union
 
 from torch import Tensor
 from torch.nn import Conv2d, init
@@ -106,8 +106,7 @@ class QConv2dBase(QConvArgsProviderMixin, QConv2d_NoAct):  # type: ignore
         return super().forward(self.activation(input_tensor))
 
 
-@QConv2dImplementation(RuntimeMode.DEFAULT)
-class QConv2d(DefaultImplementationMixin, QConv2dBase):
+class _QConv2dComposed(DefaultImplementationMixin, QConv2dBase):
     """
     This class defines the default implementation of a QConv2d layer (which is actually implemented by QConv2dBase).
 
@@ -115,3 +114,11 @@ class QConv2d(DefaultImplementationMixin, QConv2dBase):
     """
 
     pass
+
+
+QConv2d: Type[_QConv2dComposed] = QConv2dImplementation(RuntimeMode.DEFAULT)(_QConv2dComposed)  # type: ignore
+"""
+This class provides the current implementation of a QConv2d layer (which is actually implemented by :class:`QConv2dBase`).
+
+To implement a custom QConv2d implementation use :class:`QConv2dBase` as a super class instead.
+"""