+++ bitorch/models/lenet.py
@@ -1,6 +1,6 @@
 import argparse
+from typing import List
 from bitorch.layers.debug_layers import ShapePrintDebug
-from bitorch.datasets.base import BasicDataset
 from bitorch.layers import QLinear, QConv2d, QActivation
 from torch import nn
 from .base import Model
@@ -41,22 +41,32 @@ class LeNet(Model):
             ShapePrintDebug(),
             nn.Flatten(),
             QActivation(activation=input_quant_2),
-            QLinear(self.num_channels_conv * 4 * 4, self.num_fc, weight_quantization=weight_quant_2),
+            QLinear(
+                self.num_channels_conv * 4 * 4,
+                self.num_fc,
+                weight_quantization=weight_quant_2,
+            ),
             nn.BatchNorm1d(self.num_fc),
             self.activation_function(),
             nn.Linear(self.num_fc, self.num_output),
         )
         return model
 
-    def __init__(self, dataset: BasicDataset, lenet_version: int = 0) -> None:
-        """builds the model, depending on mode in either quantized or full_precision mode
+    def __init__(self, input_shape: List[int], num_classes: int = 0, lenet_version: int = 0) -> None:
+        """builds the model depending on mode in either quantized or full_precision mode
 
         Args:
-            lenet_quantized (bool, optional): toggles use of quantized version of lenet. Default is False.
+            input_shape (List[int]): input shape of images
+            num_classes (int, optional): number of output classes. Defaults to None.
+            lenet_version (int, optional): lenet version. if version outside of [0, 3], the full precision version is used. Defaults to 0.
+
+        Raises:
+            ValueError: thrown if num classes is none
         """
-        super(LeNet, self).__init__(dataset)
-        self.input_channels = dataset.shape[1]
-        self.num_output = dataset.num_classes
+        super(LeNet, self).__init__(input_shape, num_classes)
+        self.input_channels = self._input_shape[1]
+        self.num_output = self._num_classes
+
         if lenet_version == 0:
             self._model = self.generate_quant_model("sign", "sign")
         elif lenet_version == 1:
@@ -84,4 +94,4 @@ class LeNet(Model):
 
     @staticmethod
     def add_argparse_arguments(parser: argparse.ArgumentParser) -> None:
-        parser.add_argument("--lenet-version", type=int, default=0, help="choose a version of lenet")
+        parser.add_argument("--lenet-version", type=int, default=0, help="choses a verion of lenet")