+++ tests/quantizations/test_quantizations.py
@@ -8,6 +8,7 @@ from bitorch.quantizations import (
     ApproxSign,
     SteHeaviside,
     SwishSign,
+    ProgressiveSign,
     quantization_from_name,
 )
 
@@ -18,6 +19,34 @@ TEST_INPUT_DATA = [
         [-1.5, -1.0, -0.3, 0.0, 0.3, 1.0, 1.5],
         [-1.0, -1.0, -1.0, 1.0, 1.0, 1.0, 1.0],
         [1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0],
+    ),
+    (
+        ProgressiveSign(),
+        1,
+        [-1.5, -1.0, -0.3, 0.0, 0.3, 1.0, 1.5],
+        [-1.0, -1.0, -0.3, 0.0, 0.3, 1.0, 1.0],
+        [1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0],
+    ),
+    (
+        ProgressiveSign(use_global_scaling=False, initial_scale=0.2),
+        1,
+        [-1.5, -1.0, -0.3, -0.1, 0.0, 0.1, 0.3, 1.0, 1.5],
+        [-1.0, -1.0, -0.788, -0.2627, 0.0, 0.2627, 0.788, 1.0, 1.0],
+        [1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0],
+    ),
+    (
+        ProgressiveSign(use_global_scaling=False, initial_scale=0.5),
+        1,
+        [-1.5, -1.0, -0.3, -0.1, -0.05, 0.0, 0.05, 0.1, 0.3, 1.0, 1.5],
+        [-1.0, -1.0, -1.0, -1.0, -0.559, 0.0, 0.559, 1.0, 1.0, 1.0, 1.0],
+        [1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0],
+    ),
+    (
+        ProgressiveSign(use_global_scaling=False, initial_scale=0.2, custom_transform=lambda x: x),
+        1,
+        [-1.5, -1.0, -0.3, -0.1, 0.0, 0.1, 0.3, 1.0, 1.5],
+        [-1.0, -1.0, -0.375, -0.125, 0.0, 0.125, 0.375, 1.0, 1.0],
+        [1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0],
     ),
     (
         ApproxSign(),
@@ -84,6 +113,7 @@ def test_quantizations(
     assert isinstance(quantization, quantization_from_name(quantization.name))
 
     y = quantization(x)
+    assert len(y) == len(x_exp)
     assert torch.allclose(y, x_exp, atol=0.001)
     assert quantization.bit_width == bits
     with pytest.deprecated_call():